+++ core/common/src/test/kotlin/com/github/jcornaz/miop/experimental/operator/DistinctReferenceUntilChanged.kt
@@ -8,8 +8,9 @@ import kotlinx.coroutines.experimental.channels.toList
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class DistincReferenceUntilChanged : OperatorTest() {
-    override fun ReceiveChannel<Int>.operator(): ReceiveChannel<Int> =
+open class DistinctReferenceUntilChanged : OperatorTest() {
+
+    override fun <T> ReceiveChannel<T>.operator(): ReceiveChannel<T> =
         distinctReferenceUntilChanged()
 
     @Test
@@ -17,6 +18,6 @@ class DistincReferenceUntilChanged : OperatorTest() {
         val ref1 = "Hello"
         val ref2 = "world"
 
-        assertEquals(listOf(ref1, ref2, ref1, ref2), receiveChannelOf(ref1, ref2, ref2, ref2, ref1, ref2).distinctReferenceUntilChanged().toList())
+        assertEquals(listOf(ref1, ref2, ref1, ref2), receiveChannelOf(ref1, ref2, ref2, ref2, ref1, ref2).operator().toList())
     }
 }